+++ src/schema.js
@@ -1,7 +1,7 @@
 /* eslint-disable class-methods-use-this, arrow-body-style */
 
 import _ from 'lodash';
-import { parseReference } from './json-schema';
+import { parseReference, dereference } from './json-schema';
 
 export function idForDataStructure(reference) {
   return `definitions/${parseReference(reference)}`;
@@ -9,14 +9,15 @@ export function idForDataStructure(reference) {
 
 /*
  * Data Structure Generator
- * Generates a dataStructure element from a JSON schema.
+ * Generates a dataStructure element from a Swagger Schema.
  *
  * >>> const generator = new DataStructureGenerator(minimNamespace);
  * >>> const dataStructure = generator.generateDataStructure({type: 'string'});
 */
 export class DataStructureGenerator {
-  constructor(minim) {
+  constructor(minim, root) {
     this.minim = minim;
+    this.root = root;
   }
 
   // Generates a data structure element representing the given schema
@@ -170,24 +171,24 @@ export class DataStructureGenerator {
       .value();
   }
 
-  /* Validates that the given schema matches the given type
-   *
+  /**
+   * Retrieve the type from the schema
    * In the case where there is no provided type, the allOf types are matched.
+   * @param {object} schema
+   * @returns {string} type
    */
-  validateSchemaTypes(schema, type) {
-    if (schema.type === type) {
-      return true;
-    }
-
+  typeForSchema(schema) {
     if (schema.type === undefined && schema.allOf && schema.allOf.length > 0) {
-      const schemasWithoutMatchingType = schema.allOf.filter((subschema) => {
-        return !this.validateSchemaTypes(subschema, type);
-      });
+      // Try to infer type from allOf values
+      const allTypes = schema.allOf.map(this.typeForSchema);
+      const uniqueTypes = _.uniq(allTypes);
 
-      return schemasWithoutMatchingType.length === 0;
+      if (uniqueTypes.length === 1) {
+        return uniqueTypes[0];
+      }
     }
 
-    return false;
+    return schema.type;
   }
 
   // Generates an element representing the given schema
@@ -207,6 +208,7 @@ export class DataStructureGenerator {
       number: NumberElement,
       integer: NumberElement,
       null: NullElement,
+      file: StringElement,
     };
 
     if (schema.allOf && schema.allOf.length === 1 && schema.definitions &&
@@ -217,6 +219,8 @@ export class DataStructureGenerator {
       return this.generateElement(schema.allOf[0]);
     }
 
+    const type = this.typeForSchema(schema);
+
     let element;
 
     if (schema.$ref) {
@@ -226,14 +230,14 @@ export class DataStructureGenerator {
       return element;
     } else if (schema.enum) {
       element = this.generateEnum(schema);
-    } else if (schema.type === 'array') {
+    } else if (type === 'array') {
       element = this.generateArray(schema);
-    } else if (this.validateSchemaTypes(schema, 'object')) {
+    } else if (type === 'object') {
       element = this.generateObject(schema);
-    } else if (schema.type && typeGeneratorMap[schema.type]) {
-      element = new typeGeneratorMap[schema.type]();
-    } else if (_.isArray(schema.type)) {
-      // TODO: Support multiple `type`
+    } else if (type && typeGeneratorMap[type]) {
+      element = new typeGeneratorMap[type]();
+    } else if (type) {
+      throw new Error(`Unhandled schema type '${type}'`);
     }
 
     if (element) {
@@ -264,10 +268,8 @@ export class DataStructureGenerator {
 
       let samples = [];
 
-      if (schema.examples) {
-        samples = schema.examples;
-      } else if (schema.example) {
-        samples = [schema.example];
+      if (schema.example) {
+        samples = [dereference(schema.example, this.root)];
       }
 
       if (samples.length > 0) {